+++ tests/cli/set/test_sample.py
@@ -4,6 +4,7 @@ from datetime import datetime
 import pytest
 from cg.cli.set import sample
 from cg.store import Store
+from tests.store_helpers import add_sample, ensure_application_version, ensure_customer
 
 SUCCESS = 0
 
@@ -22,7 +23,7 @@ def test_invalid_sample(cli_runner, base_context):
 def test_skip_lims(cli_runner, base_context, base_store: Store):
     # GIVEN a database with a sample
 
-    sample_obj = add_sample(base_store, sex="female")
+    sample_obj = add_sample(base_store, gender="female")
     key = "name"
     new_value = "new_value"
 
@@ -43,7 +44,7 @@ def test_skip_lims(cli_runner, base_context, base_store: Store):
 def test_set_sample(cli_runner, base_context, base_store: Store, key):
     # GIVEN a database with a sample
 
-    sample_obj = add_sample(base_store, sex="female")
+    sample_obj = add_sample(base_store, gender="female")
     new_value = "new_value"
     assert getattr(sample_obj, key) != new_value
 
@@ -62,7 +63,7 @@ def test_set_sample(cli_runner, base_context, base_store: Store, key):
 def test_sex(cli_runner, base_context, base_store: Store):
     # GIVEN a database with a sample
 
-    sample_obj = add_sample(base_store, sex="female")
+    sample_obj = add_sample(base_store, gender="female")
     key = "sex"
     new_value = "male"
     assert getattr(sample_obj, key) != new_value
@@ -87,9 +88,7 @@ def test_invalid_customer(cli_runner, base_context, base_store: Store):
 
     # WHEN calling set sample with an invalid customer
     result = cli_runner.invoke(
-        sample,
-        [sample_id, "-kv", "customer", customer_id, "-y", "--skip-lims"],
-        obj=base_context,
+        sample, [sample_id, "-kv", "customer", customer_id, "-y", "--skip-lims"], obj=base_context
     )
 
     # THEN then it should error about missing customer instead of setting the value
@@ -105,9 +104,7 @@ def test_customer(cli_runner, base_context, base_store: Store):
 
     # WHEN calling set sample with a valid customer
     result = cli_runner.invoke(
-        sample,
-        [sample_id, "-kv", "customer", customer_id, "-y", "--skip-lims"],
-        obj=base_context,
+        sample, [sample_id, "-kv", "customer", customer_id, "-y", "--skip-lims"], obj=base_context
     )
 
     # THEN then it should set the customer of the sample
@@ -121,9 +118,7 @@ def test_invalid_downsampled_to(cli_runner, base_context, base_store: Store):
 
     # WHEN calling set sample with an invalid value of downsampled to
     result = cli_runner.invoke(
-        sample,
-        ["dummy_sample_id", "-kv", "downsampled_to", downsampled_to, "-y"],
-        obj=base_context,
+        sample, ["dummy_sample_id", "-kv", "downsampled_to", downsampled_to, "-y"], obj=base_context
     )
 
     # THEN wrong data type
@@ -138,9 +133,7 @@ def test_downsampled_to(cli_runner, base_context, base_store: Store):
 
     # WHEN calling set sample with a valid value of downsampled to
     result = cli_runner.invoke(
-        sample,
-        [sample_id, "-kv", "downsampled_to", downsampled_to, "-y"],
-        obj=base_context,
+        sample, [sample_id, "-kv", "downsampled_to", downsampled_to, "-y"], obj=base_context
     )
 
     # THEN then the value should have been set on the sample
@@ -168,10 +161,7 @@ def test_invalid_application(cli_runner, base_context, base_store: Store):
     # GIVEN a database with a sample
     sample_id = add_sample(base_store).internal_id
     application_tag = "dummy_application"
-    assert (
-        base_store.Sample.query.first().application_version.application.tag
-        != application_tag
-    )
+    assert base_store.Sample.query.first().application_version.application.tag != application_tag
 
     # WHEN calling set sample with an invalid application
     result = cli_runner.invoke(
@@ -182,22 +172,14 @@ def test_invalid_application(cli_runner, base_context, base_store: Store):
 
     # THEN then it should error about missing application instead of setting the value
     assert result.exit_code != SUCCESS
-    assert (
-        base_store.Sample.query.first().application_version.application.tag
-        != application_tag
-    )
+    assert base_store.Sample.query.first().application_version.application.tag != application_tag
 
 
 def test_application(cli_runner, base_context, base_store: Store):
     # GIVEN a database with a sample and two applications
     sample_obj = add_sample(base_store)
-    application_tag = ensure_application_version(
-        base_store, "another_application"
-    ).application.tag
-    assert (
-        base_store.Sample.query.first().application_version.application.tag
-        != application_tag
-    )
+    application_tag = ensure_application_version(base_store, "another_application").application.tag
+    assert base_store.Sample.query.first().application_version.application.tag != application_tag
 
     # WHEN calling set sample with an invalid application
     result = cli_runner.invoke(
@@ -216,57 +198,3 @@ def test_application(cli_runner, base_context, base_store: Store):
     # THEN then the application should have been set
     assert result.exit_code == SUCCESS
     assert sample_obj.application_version.application.tag == application_tag
-
-
-def ensure_application_version(store, application_tag="dummy_tag"):
-    """utility function to return existing or create application version for tests"""
-    application = store.application(tag=application_tag)
-    if not application:
-        application = store.add_application(
-            tag=application_tag,
-            category="wgs",
-            description="dummy_description",
-            percent_kth=80,
-        )
-        store.add_commit(application)
-
-    prices = {"standard": 10, "priority": 20, "express": 30, "research": 5}
-    version = store.application_version(application, 1)
-    if not version:
-        version = store.add_version(
-            application, 1, valid_from=datetime.now(), prices=prices
-        )
-
-        store.add_commit(version)
-    return version
-
-
-def ensure_customer(store, customer_id="cust_test"):
-    """utility function to return existing or create customer for tests"""
-    customer_group_id = customer_id + "_group"
-    customer_group = store.customer_group(customer_group_id)
-    if not customer_group:
-        customer_group = store.add_customer_group(customer_group_id, customer_group_id)
-
-        customer = store.add_customer(
-            internal_id=customer_id,
-            name="Test Customer",
-            scout_access=False,
-            customer_group=customer_group,
-            invoice_address="dummy_address",
-            invoice_reference="dummy_reference",
-        )
-        store.add_commit(customer)
-    customer = store.customer(customer_id)
-    return customer
-
-
-def add_sample(store, sample_id="sample_test", sex="female"):
-    """utility function to add a sample to use in tests"""
-    customer = ensure_customer(store)
-    application_version_id = ensure_application_version(store).id
-    sample_obj = store.add_sample(name=sample_id, sex=sex)
-    sample_obj.application_version_id = application_version_id
-    sample_obj.customer = customer
-    store.add_commit(sample_obj)
-    return sample_obj
